refactorings - learn_skill from unit.c to study.c - add academy.c module
