+++ src/study.c
@@ -25,6 +25,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "move.h"
 #include "monster.h"
 #include "alchemy.h"
+#include "academy.h"
 
 #include <kernel/ally.h>
 #include <kernel/building.h>
@@ -48,6 +49,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <util/log.h>
 #include <util/parser.h>
 #include <util/rand.h>
+#include <util/rng.h>
 #include <util/umlaut.h>
 
 /* libc includes */
@@ -57,8 +59,6 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
-
-#define TEACHNUMBER 10
 
 static skill_t getskill(const struct locale *lang)
 {
@@ -484,9 +484,8 @@ int teach_cmd(unit * u, struct order *ord)
         replace_order(&u->orders, ord, new_order);
         free_order(new_order);      /* parse_order & set_order have each increased the refcount */
     }
-    if (academy && sk != NOSKILL) {
-        academy = academy / 30;     /* anzahl gelehrter wochen, max. 10 */
-        learn_skill(u, sk, academy / 30.0 / TEACHNUMBER);
+    if (academy) {
+        academy_teaching_bonus(u, sk, academy);
     }
     return 0;
 }
@@ -823,4 +822,28 @@ void produceexp_ex(struct unit *u, skill_t sk, int n, bool(*learn)(unit *, skill
 void produceexp(struct unit *u, skill_t sk, int n)
 {
     produceexp_ex(u, sk, n, learn_skill);
+}
+
+bool learn_skill(unit * u, skill_t sk, double learn_chance)
+{
+    skill *sv = u->skills;
+    if (learn_chance < 1.0 && rng_int() % 10000 >= learn_chance * 10000)
+        if (!chance(learn_chance))
+            return false;
+    while (sv != u->skills + u->skill_size) {
+        assert(sv->weeks > 0);
+        if (sv->id == sk) {
+            if (sv->weeks <= 1) {
+                sk_set(sv, sv->level + 1);
+            }
+            else {
+                sv->weeks--;
+            }
+            return true;
+        }
+        ++sv;
+    }
+    sv = add_skill(u, sk);
+    sk_set(sv, 1);
+    return true;
 }
